Implemented a aggregate extension to calculate medians
Implemented an aggregate extension that calculates the median of the
values passed to it. The extension collects all the data over the
transition function calls and then calculates the median after sorting
all the collected values in the final function call. This patch provides
support for calculating medians for smallint and int data types.